requirements.txt
astroid==2.3.3
autopep8==1.4.4
cairocffi==1.0.0
CairoSVG==2.3.0
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
Click==7.0
colorama==0.4.1
coreapi==2.3.3
coreschema==0.0.4
cssselect2==0.2.1
defusedxml==0.5.0
diff-match-patch==20121119
Django==2.2
django-cors-headers==2.2.0
django-easy-pdf==0.1.1
django-filter==2.0.0
django-import-export==1.0.1
django-jalali==3.0.3
django-oauth-plus==2.2.9
django-reset-migrations==0.3.1
django-rest-swagger==2.2.0
django-stubs==1.4.0
django-wkhtmltopdf==3.2.0
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
djangorestframework-oauth==1.1.0
djangorestframework-recursive==0.1.2
djangorestframework-stubs==1.0.0
docutils==0.14
drf-yasg==1.16.1
et-xmlfile==1.0.1
Faker==2.0.0
Flask==1.1.1
html5lib==1.0.1
httplib2==0.17.0
idna==2.6
inflection==0.3.1
isort==4.3.21
itsdangerous==1.1.0
itypes==1.1.0
jdatetime==3.0.0
jdcal==1.4
Jinja2==2.10.1
lazy-object-proxy==1.4.3
lml==0.0.4
MarkupSafe==1.1.1
mccabe==0.6.1
mypy==0.720
mypy-extensions==0.4.3
oauth2==1.9.0.post1
odfpy==1.3.6
openapi-codec==1.3.2
openpyxl==2.5.7
Pillow==7.0.0
psycopg2==2.8.2
pycodestyle==2.5.0
pycparser==2.19
pyexcel==0.5.9.1
pyexcel-io==0.5.9.1
pyexcel-webio==0.1.4
PyJWT==1.6.4
pylint==2.4.4
pymongo==3.9.0
PyPDF2==1.26.0
Pyphen==0.9.5
pyserial==3.4
python-dateutil==2.8.0
pytz==2018.4
PyYAML==3.13
reportlab==3.5.13
requests==2.18.4
ruamel.yaml==0.15.100
simplejson==3.16.0
six==1.12.0
sqlparse==0.3.0
tablib==0.12.1
text-unidecode==1.2
texttable==1.4.0
tinycss2==0.6.1
typed-ast==1.4.0
typing-extensions==3.7.4.1
unicodecsv==0.14.1
uritemplate==3.0.0
urllib3==1.22
virtualenv==16.6.0
virtualenvwrapper-win==1.2.5
WeasyPrint==44
webencodings==0.5.1
Werkzeug==0.16.0
wrapt==1.11.2
xhtml2pdf==0.2.3
xlrd==1.1.0
xlwt==1.3.0
